+++ tests/Unit/RequestMacroTest.php
@@ -4,25 +4,9 @@ namespace Bilfeldt\RequestLogger\Tests\Unit;
 
 use Bilfeldt\RequestLogger\Tests\TestCase;
 use Illuminate\Http\Request;
-use Illuminate\Support\Str;
 
 class RequestMacroTest extends TestCase
 {
-    public function test_request_macro_get_unique_id_is_registered()
-    {
-        $this->assertTrue((new Request())->hasMacro('getUniqueId'));
-    }
-
-    public function test_request_macro_get_unique_id()
-    {
-        $request = new Request();
-
-        $uuid = $request->getUniqueId();
-
-        $this->assertTrue(Str::isUuid($uuid));
-        $this->assertEquals($uuid, $request->getUniqueId());
-    }
-
     public function test_request_macro_enable_log_registered()
     {
         $this->assertTrue((new Request())->hasMacro('enableLog'));